% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_connection}
\alias{get_connection}
\title{Returns a jdbc connection to redshift}
\usage{
get_connection(connection_string)
}
\arguments{
\item{connection_string}{(chr) connection string}
}
\value{
jdbc connection
}
\description{
Returns an active jdbc connection given a connection_string
}
