% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_environment_variable}
\alias{get_environment_variable}
\title{Get environment wariable with additional warning handling}
\usage{
get_environment_variable(variable_name, valid_values = NULL)
}
\arguments{
\item{variable_name}{(chr) variable name}

\item{valid_values}{vector or list. acceptable values for an environment variable}
}
\value{
(chr) Contents of environment variable
}
\description{
Wraps `Sys.getenv` to get and validate an environment variable.
}
